web/requirements.txt
Flask>=1.1.1
pytz
pyyaml
s3path
boto3
minio
python-dotenv==0.12.0
gunicorn==20.0.4
corems>=1.0.0
Flask-Login==0.5.0
flask-sqlalchemy
Flask-Migrate==2.5.3
bokeh==2.0.2
flask-wtf>=0.14.3
wtforms[email]==2.3.1
celery[redis]
psycopg2-binary==2.8.4
natsort==7.0.1
colorcet==2.0.2
Flask-Markdown>=0.3
pyjwt
requests
